fix(phpstan): improve compatibility with quiqqer/erp
Added checks for the existence of various classes from quiqqer/erp before trying to use them. This
ensures that the code fails gracefully if the expected classes are not present, for instance if the
quiqqer/erp package is not installed.

Also added a new dependency on quiqqer/erp in `composer.json`. This will prevent the code from
being installed without the necessary package, but in case it's removed or not loaded for some
reason, the new checks still provide a safety net.